requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-debug-toolbar==4.3.0
django-jalali-date==1.1.0
django-widget-tweaks==1.5.0
idna==3.7
jalali_core==1.0.0
jdatetime==5.0.0
kavenegar==1.1.2
mysqlclient==2.2.4
pillow==10.3.0
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
